move  test chain to chain mock
